+++ NEWS.md
@@ -11,7 +11,7 @@
 * A `%->%` operator has been added. The right operator performs the same
   operation as `%<-%` with the name structure on the right-hand side and
   the values to assign on the left-hand side.
-* `<-` may be used to specify the default value of a variable. A default value
+* `=` may be used to specify the default value of a variable. A default value
   is used when there are an insufficient number of values.
 
 # zeallot 0.0.4